+++ CHANGELOG.md
@@ -7,7 +7,16 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Breaking Changes
 
-- None
+- [#1108](https://github.com/paper-trail-gem/paper_trail/pull/1108) -
+  In `versions.item_type`, we now store the subclass name instead of
+  the base_class.
+  - You must migrate existing `versions` records if you use
+    [STI][1]. A migration generator has been provided. Generator `update_sti`
+    creates a migration that updates existing `version` entries such that
+    `item_type` then refers to the specific class name instead of base_class.
+    See [5.c. Generators][2] for instructions.
+  - This change fixes a long-standing issue with reification of STI subclasses,
+    [#594](https://github.com/paper-trail-gem/paper_trail/issues/594)
 
 ### Added
 
@@ -1022,3 +1031,6 @@ in the `PaperTrail::Version` class through a `Rails::Engine` when the gem is use
   - [#160](https://github.com/paper-trail-gem/paper_trail/pull/160) - Fixed failing tests and resolved out of date dependency issues.
   - [#157](https://github.com/paper-trail-gem/paper_trail/pull/157) - Refactored `class_attribute` names on the `ClassMethods` module
     for names that are not obviously pertaining to PaperTrail to prevent method name collision.
+
+[1]: https://api.rubyonrails.org/v5.2.0/classes/ActiveRecord/Base.html#class-ActiveRecord::Base-label-Single+table+inheritance
+[2]: https://github.com/paper-trail-gem/paper_trail/blob/master/README.md#5c-generators